Working code
Renamed article to citation. Added ViewModels and a model mock
(currently not unit tested). Currently the code can add files in a list
of files and can search for a folder and store it. I added some clumsy
hard-coded instructions in the UI, will have to move that in a pop-up
window or something more practical.